create

+++ Increasing_Array.cpp
@@ -20,7 +20,7 @@ int main()
 {
      int n;
     cin >> n;
-    vector<int> v(n);
+    vector<long long> v(n);
     for (int i = 0; i < n; i++)
         cin >> v[i];
     long long ans = 0;
